Add master password strength checks to `PwdManager`
Dropbox's `zxcvbn` password strength estimator is used to help ensure
that the master password used for the password manager is strong. The
manager enforces that the master password strength estimate is the
maximum possible score of 4. If the password strength is less than 4,
functions that set or update the master password return the newly
introduced error variant, `WeakPassword`.