Add `found_signature()` method to `PwdManager` (#7)
A unique 'file signature' specific to `pwdm` has been introduced. The
manager now cross-verifies this signature whenever a database connection
is successfully opened, essentially confirming that the database has been
originally configured by `pwdm`. This helps, for example, preventing any
inadvertent operations on databases that are unknown to or have not been
configured by `pwdm`. The 'file signature' is stored in the database.