Use service and username to identify each password (#15)
Refactored the password manager to utilize a service-username pair
(represented by the `UserIdentity` struct), rather than an ID string, to
identify each password. This is more aligned with real-world user
behavior, as users tend to associate a password with a username and a
particular target platform, not with an arbitrary ID. This change makes
the password manager easier to use.