Fix formatting issues in CLI "List" table (#16)
This commit ensures that the border line (composed of '=' characters)
under the header of the "List" table in the CLI is not shorter than the
header text. The border line would be shorter, for example, if there is
only a single entry with no username or a short username. Moreover,
it ensures that the '|' separator for a (service, username) entry always
aligns with the header's '|'.